RAGs/Reflection_RAG/requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pypdf>=3.0.0
